+++ ReleaseNotes.md
@@ -11,3 +11,4 @@
 
 - Introduces tags for instances
 - Introduces support for instance filtering when deploying (using `IServiceFilter`)
+- Introduces new `Deploys(IServiceManager)` extension to simplify Task declarations